writer/requirements.txt
# Frank Writer Requirements

# GTK4 / UI (system packages, installed via apt)
# PyGObject>=3.44.0
# pycairo>=1.24.0

# Text Processing
cmarkgfm>=2024.1.0
Pygments>=2.17.0

# Document Formats
python-docx>=1.1.0
docxtpl>=0.16.0
PyPDF2>=3.0.0

# PDF Generation
WeasyPrint>=60.0

# Templates
Jinja2>=3.1.0
PyYAML>=6.0.0

# Async/HTTP
httpx>=0.25.0
aiofiles>=23.0.0
